@@ -10,27 +10,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.2.10",
+    "@angular/animations": "^16.2.12",
     "@angular/cdk": "^15.2.9",
-    "@angular/common": "^15.1.0",
-    "@angular/compiler": "^15.1.0",
-    "@angular/core": "^15.1.0",
-    "@angular/forms": "^15.1.0",
+    "@angular/common": "^16.2.12",
+    "@angular/compiler": "^16.2.12",
+    "@angular/core": "^16.2.12",
+    "@angular/forms": "^16.2.12",
     "@angular/material": "^15.2.9",
-    "@angular/platform-browser": "^15.1.0",
-    "@angular/platform-browser-dynamic": "^15.1.0",
-    "@angular/router": "^15.1.0",
+    "@angular/platform-browser": "^16.2.12",
+    "@angular/platform-browser-dynamic": "^16.2.12",
+    "@angular/router": "^16.2.12",
     "dotenv": "^16.4.5",
     "ngx-cookie-service": "^17.0.0",
     "posthog-js": "^1.118.1",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.12.0"
+    "zone.js": "~0.13.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.1.5",
-    "@angular/cli": "~15.1.5",
-    "@angular/compiler-cli": "^15.1.0",
+    "@angular-devkit/build-angular": "^16.2.13",
+    "@angular/cli": "~16.2.13",
+    "@angular/compiler-cli": "^16.2.12",
     "@types/jasmine": "~4.3.0",
     "@types/node": "^20.12.5",
     "jasmine-core": "~4.5.0",
